Removed parallel lifetime of ModelDescriptor - lifetime is now again bound to MojoModel. Its construction is mandatory.
