+++ h2o-genmodel/src/main/java/hex/genmodel/MojoModel.java
@@ -9,7 +9,6 @@ import java.io.*;
 public abstract class MojoModel extends GenModel {
 
   public String _h2oVersion;
-  public String algoName;
   public hex.ModelCategory _category;
   public String _uuid;
   public boolean _supervised;
@@ -20,6 +19,7 @@ public abstract class MojoModel extends GenModel {
   public double[] _priorClassDistrib;
   public double[] _modelClassDistrib;
   public double _mojo_version;
+  public ModelDescriptor _modelDescriptor = null;
 
   /**
    * Primary factory method for constructing MojoModel instances.
