+++ h2o-genmodel/src/main/java/hex/genmodel/algos/tree/SharedTreeMojoReader.java
@@ -4,6 +4,8 @@ import com.google.gson.JsonArray;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 import hex.genmodel.ModelMojoReader;
+import hex.genmodel.descriptor.JsonModelDescriptorReader;
+import hex.genmodel.descriptor.ModelDescriptorBuilder;
 import hex.genmodel.descriptor.Table;
 import hex.genmodel.descriptor.VariableImportances;
 
@@ -54,5 +56,23 @@ public abstract class SharedTreeMojoReader<M extends SharedTreeMojoModel> extend
 
 
     _model.postInit();
+  }
+
+  @Override
+  protected void readModelSpecificDescription(final ModelDescriptorBuilder modelDescriptorBuilder, final JsonObject modelJson) {
+    modelDescriptorBuilder.variableImportances(extractVariableImportances(modelJson));
+  }
+
+  private VariableImportances extractVariableImportances(final JsonObject modelJson) {
+    final Table table = JsonModelDescriptorReader.extractTableFromJson(modelJson, "output.variable_importances");
+    if (table == null) return null;
+    final double[] relativeVarimps = new double[table.rows()];
+    final int column = table.findColumnIndex("Relative Importance");
+    if (column == -1) return null;
+    for (int i = 0; i < table.rows(); i++) {
+      relativeVarimps[i] = (double) table.getCell(column, i);
+    }
+
+    return new VariableImportances(Arrays.copyOf(_model._names, _model.nfeatures()), relativeVarimps);
   }
 }
